combine dashboard with username page
