+++ src/lib/components/ProfileButton.svelte
@@ -2,7 +2,9 @@
     import { goto } from "$app/navigation";
 </script>
 
-<button class="h-4vh w-4vh mr-2vh" on:click={() => goto("/username")}><img src="icons/iconoir--profile-circle.svg"></button>
+<button class="h-4vh w-4vh mr-2vh" on:click={() => goto("/username")}>
+    <img src="icons/iconoir--profile-circle.svg">
+</button>
 
 <style>
